Add a transformer for static files and use it for css, js, …
Instead of having a transformer with a hardcoded css path and one for
the favicon, we now have a general static asset transformer that finds
a file for any path that starts with '/static/'.

This allows us to pull in static assets from CDNs and to host them
locally — we can now run in offline mode.